feat(templates): Rename badger helper folder to common
