+++ src/apps/badger/arbitrum/badger.vault.token-fetcher.ts
@@ -2,7 +2,7 @@ import { Register } from '~app-toolkit/decorators';
 import { Network } from '~types/network.interface';
 
 import { BADGER_DEFINITION } from '../badger.definition';
-import { BadgerVaultTokenFetcher } from '../helpers/badger.vault.token-fetcher';
+import { BadgerVaultTokenFetcher } from '../common/badger.vault.token-fetcher';
 
 const appId = BADGER_DEFINITION.id;
 const groupId = BADGER_DEFINITION.groups.vault.id;
