+++ src/apps/badger/badger.module.ts
@@ -4,11 +4,11 @@ import { AbstractApp } from '~app/app.dynamic-module';
 import { ArbitrumBadgerClaimableContractPositionFetcher } from './arbitrum/badger.claimable.contract-position-fetcher';
 import { ArbitrumBadgerVaultTokenFetcher } from './arbitrum/badger.vault.token-fetcher';
 import { BadgerAppDefinition, BADGER_DEFINITION } from './badger.definition';
+import { BadgerClaimableRewardsResolver } from './common/badger.claimable.rewards-resolver';
+import { BadgerVaultTokenDefinitionsResolver } from './common/badger.vault.token-definition-resolver';
 import { BadgerContractFactory } from './contracts';
 import { EthereumBadgerClaimableContractPositionFetcher } from './ethereum/badger.claimable.contract-position-fetcher';
 import { EthereumBadgerVaultTokenFetcher } from './ethereum/badger.vault.token-fetcher';
-import { BadgerClaimableRewardsResolver } from './helpers/badger.claimable.rewards-resolver';
-import { BadgerVaultTokenDefinitionsResolver } from './helpers/badger.vault.token-definition-resolver';
 import { PolygonBadgerClaimableContractPositionFetcher } from './polygon/badger.claimable.contract-position-fetcher';
 import { PolygonBadgerVaultTokenFetcher } from './polygon/badger.vault.token-fetcher';
